rename model to record
